BA-002: deps to latest stable
Cargo.toml:
- actix-web 4.7 -> 4.13
- rusqlite 0.31 -> 0.40
- chrono 0.4.38 -> 0.4.44
- serde 1.0 -> 1.0.228 (with derive feature)
- serde_json 1.0 -> 1.0.150
- mockall 0.12 -> 0.14, rstest 0.21 -> 0.26 (both dev)
- drop serde_derive (replaced by serde derive feature)
- drop dotenv (unmaintained) -> dotenvy 0.15.7
- toolchain bumped to rustc 1.95 (rusqlite 0.40 needs cfg_select)

Source adjustments forced by upgrades:
- u64 -> i64 for SQLite timestamp columns: rusqlite 0.40 dropped
  ToSql/FromSql for u64 (SQLite INTEGER is i64). Affects Game.date,
  MatchInfo.date, DbGame.utc_date, DbTip.date, plus the now() cast
  in fixtures::load_fixtures. Wire-level JSON output unchanged
  (still a number).
- sort_by(|a, b| b.x.cmp(&a.x)) -> sort_by_key(|x| Reverse(x))
  (clippy 1.95 unnecessary_sort_by) in routes.rs::user_by_id and
  service::calculate_positions
- use serde_derive -> use serde, use dotenv -> use dotenvy

cargo clippy --all-targets -- -D warnings: green
MODE=test cargo test: 72/72 passed

+++ src/db/fixtures.rs
@@ -20,7 +20,7 @@ struct DbGame {
     home_team: String,
     away_team: String,
     status: String,
-    utc_date: u64,
+    utc_date: i64,
     home_score: Option<i32>,
     away_score: Option<i32>,
 }
@@ -29,7 +29,7 @@ struct DbGame {
 struct DbTip {
     user_id: i32,
     match_id: i32,
-    date: u64,
+    date: i64,
     score_home: i32,
     score_away: i32,
 }
@@ -90,7 +90,7 @@ pub fn load_fixtures(conn: &Connection) {
     let now = SystemTime::now()
         .duration_since(UNIX_EPOCH)
         .unwrap()
-        .as_secs();
+        .as_secs() as i64;
 
     let games = get_games(&lands, now);
 
@@ -102,7 +102,7 @@ pub fn load_fixtures(conn: &Connection) {
     insert_tips(conn, &tips).unwrap();
 }
 
-fn get_tips(now: u64) -> Vec<DbTip> {
+fn get_tips(now: i64) -> Vec<DbTip> {
     let tpis = vec![
         DbTip {
             user_id: 1,
@@ -185,7 +185,7 @@ fn get_tips(now: u64) -> Vec<DbTip> {
     tpis
 }
 
-fn get_games(lands: &HashMap<&str, Team>, now: u64) -> Vec<DbGame> {
+fn get_games(lands: &HashMap<&str, Team>, now: i64) -> Vec<DbGame> {
     let games = vec![
         DbGame {
             id: 1,
